so much stuff

+++ interfaces/src/main/java/ar/edu/itba/paw/interfaces/Dao.java
@@ -3,6 +3,6 @@ package ar.edu.itba.paw.interfaces;
 import java.util.Collection;
 
 public interface Dao<T> {
-    T get(int id);
+    T get(long id);
     Collection<T> getAll();
 }
